Moving new notifications endpoint and making it a little bit more unified with other endpoints
